102 - SMTP configuration not used
The SMTP configration is being added to ActionMailer with string keys
but need to be sybmolized keys. As a result, at run time the
configured SMTP settings (e.g. from
config/environments/common/mailer.yml) are ignored and the defaults are
used instead.

This symbolizes the keys that are set on
ActionMailer::Base.smtp_settings